+++ src/main/java/com/tuannh/phantom/db/internal/file/TombstoneFile.java
@@ -28,8 +28,8 @@ public class TombstoneFile implements Closeable {
     private final File file;
     private final FileChannel channel;
     //
-    private long unflushed = 0;
-    private long writeOffset = 0;
+    private int unflushed = 0;
+    private int writeOffset = 0;
 
     private TombstoneFile(int fileId, DBDirectory dbDirectory, PhantomDBOptions dbOptions, File file, FileChannel channel) {
         this.fileId = fileId;
